+++ templates/AttachActor.hbs
@@ -1,0 +1,19 @@
+<h1>Assign one actor at a time to one or multiple tokens.</h1>
+<h2>
+    <p>If there is nothing special about your token, then just replace it with a token dragged to the canvas from an actor.</p>
+    <p>Only resort to this if your token has unique values you wish to preserve!</p>
+</h2>
+<div>
+    <table>
+        <tr>
+            <td style="width: 200px"><label for="attachActor">Actor must be in Actor's folder... Type actor name here: </label></td>
+            <td style="text-align: left"><input type="text" id="attachActor" name="attachActor" style="width: 200px"></td>
+        </tr>
+        {{#each tokens}}
+            <tr>
+                <td style="width: 30px"><input type="checkbox" id="check" name="{{this.name}}" value="{{this.id}}"></td>
+                <td style="text-align: left"><label for="{{this.name}}">{{this.name}} ({{this.tokenName}})</label>
+            </tr>
+        {{/each}}
+    </table>
+</div>
